pretty good

+++ view/displayuser.php
@@ -17,18 +17,19 @@ if (isset($data)) {
             if ($k == 'lastname'){
                 $lastname = $v;
                 echo "Lastname : $lastname&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp";
-            } 
-            if ($k == 'password'){
-                $password = $v;
-                echo "Password : $password&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp";
             }
             if ($k == 'role'){
                 $role = $v;
                 echo "Role : $role</div>";
             }
         }
-    echo "<ul><a href='edituser.php/?username=$username'>Edit</a>&nbsp;&nbsp;";
-    echo "<a href='../controller/deleteusercontroller.php/?username=$username'>Delete</a></ul></div></ul>";
+        if($_SESSION['role'] == 'librarian'){
+            echo "<ul><a href='edituser.php/?username=$username'>Edit</a>&nbsp;&nbsp;";
+            echo "<a href='../controller/deleteusercontroller.php/?username=$username'>Delete</a></ul></div></ul>";
+        }else{
+            echo "</div>";
+        }
+        echo "</div>";
     }
 }
 include "footerv.php";
